Apply changes to accommodate Cucumber tests
Make routes having just 1 fragment after the base URL.
Adopt conventional HTML identifier for radio buttons.
