+++ app/views/lettingHistory/occupierList.scala.html
@@ -91,6 +91,7 @@
         @includes.radioButtonsYesNo(
             govukRadios,
             theForm,
+            idPrefix = "lettingHistoryOccupierList",
             name = "answer",
             legend = messages(
                 "lettingHistory.occupiersList.hadMoreOccupiers.label",
