+++ app/views/lettingHistory/residentList.scala.html
@@ -78,6 +78,7 @@
         @includes.radioButtonsYesNo(
             govukRadios,
             theForm,
+            idPrefix = "lettingHistoryResidentList",
             name = "answer",
             legendKey = "lettingHistory.residentList.hasMoreResidents.label",
             classes = "govuk-fieldset__legend--s",
