+++ test/controllers/lettingHistory/HasPermanentResidentsControllerSpec.scala
@@ -26,113 +26,105 @@ import play.api.libs.json.Writes
 import play.api.mvc.Codec.utf_8 as UTF_8
 import play.api.test.Helpers._
 import uk.gov.hmrc.http.HeaderCarrier
-import views.html.lettingHistory.completedLettings as CompletedLettingsView
+import views.html.lettingHistory.hasPermanentResidents as HasPermanentResidentsView
 
-class CompletedLettingsControllerSpec extends LettingHistoryControllerSpec:
+class HasPermanentResidentsControllerSpec extends LettingHistoryControllerSpec:
 
-  "the CompletedLettings controller" when {
+  "the PermanentResidents controller" when {
     "the user has not provided any answer yet" should {
-      "be handling GET by replying 200 with the HTML form having unchecked radios" in new ControllerFixture {
+      "be handling GET and reply 200 with the HTML form having unchecked radios" in new ControllerFixture {
         val result = controller.show(fakeGetRequest)
         status(result)            shouldBe OK
         contentType(result).value shouldBe HTML
         charset(result).value     shouldBe UTF_8.charset
         val page = contentAsJsoup(result)
-        page.heading        shouldBe "lettingHistory.completedLettings.heading"
-        page.backLink       shouldBe routes.PermanentResidentsController.show.url
+        page.heading        shouldBe "lettingHistory.permanentResidents.heading"
+        page.backLink       shouldBe controllers.routes.TaskListController.show().withFragment("lettingHistory").toString
         page.radios("answer") should haveNoneChecked
       }
       "be handling invalid POST by replying 400 with error message" in new ControllerFixture {
         val result = controller.submit(
           fakePostRequest.withFormUrlEncodedBody(
-            "answer" -> ""
+            "answer" -> "" // missing
           )
         )
         status(result) shouldBe BAD_REQUEST
         val page   = contentAsJsoup(result)
-        page.error("answer") shouldBe "lettingHistory.hasCompletedLettings.required"
+        page.error("answer") shouldBe "lettingHistory.hasPermanentResidents.required"
       }
-      "be handling POST answer='yes' by replying 303 redirect to 'Occupier Detail' page" in new ControllerFixture {
+      "be handling POST answer='yes' by replying 303 redirect to the 'Residents Details' page" in new ControllerFixture {
         val result = controller.submit(
           fakePostRequest.withFormUrlEncodedBody(
             "answer" -> "yes"
           )
         )
         status(result) shouldBe SEE_OTHER
-        redirectLocation(result).value   shouldBe routes.OccupierDetailController.show(index = None).url
+        redirectLocation(result).value    shouldBe routes.ResidentDetailController.show().url
         verify(repository, once).saveOrUpdate(data.capture())(any[Writes[Session]], any[HeaderCarrier])
-        hasCompletedLettings(data).value shouldBe true
+        hasPermanentResidents(data).value shouldBe true
       }
     }
-    "the user has already provided an answer"  should {
-      "regardless of the given number of occupiers"          should {
-        "be handling GET by replying 200 with the HTML form having checked radios" in new ControllerFixture(
-          permanentResidents = twoResidents,
-          hasCompletedLettings = Some(true)
+    "the user has already answered"            should {
+      "regardless of the given number of residents"          should {
+        "be handling GET and reply 200 with the HTML form having checked radios" in new ControllerFixture(
+          oneResident
         ) {
           val result = controller.show(fakeGetRequest)
           status(result)            shouldBe OK
           contentType(result).value shouldBe HTML
           charset(result).value     shouldBe UTF_8.charset
           val page = contentAsJsoup(result)
-          page.backLink          shouldBe routes.ResidentListController.show.url
-          page.radios("answer")    should haveChecked(value = "yes")
-          page.radios("answer") shouldNot haveChecked(value = "no")
+          page.radios("answer")    should haveChecked("yes")
+          page.radios("answer") shouldNot haveChecked("no")
         }
-        "be handling POST answer='yes' by replying 303 redirect to the 'Occupier Detail' page" in new ControllerFixture(
-          hasCompletedLettings = Some(true)
+        "be handling POST answer='yes' by replying 303 redirect to the 'Resident Detail' page" in new ControllerFixture(
+          oneResident
         ) {
           val result = controller.submit(fakePostRequest.withFormUrlEncodedBody("answer" -> "yes"))
-          status(result)                   shouldBe SEE_OTHER
-          redirectLocation(result).value   shouldBe routes.OccupierDetailController.show(index = None).url
+          status(result)                    shouldBe SEE_OTHER
+          redirectLocation(result).value    shouldBe routes.ResidentDetailController.show().url
           verify(repository, once).saveOrUpdate(data.capture())(any[Writes[Session]], any[HeaderCarrier])
-          hasCompletedLettings(data).value shouldBe true
+          hasPermanentResidents(data).value shouldBe true
         }
-        "be handling POST answer='no' by replying 303 redirect to the 'Letting intention' page" in new ControllerFixture(
-          hasCompletedLettings = Some(true)
+        "be handling POST answer='no' by replying 303 redirect to the 'Commercial Lettings' page" in new ControllerFixture(
+          oneResident
         ) {
           // Answering 'no' will clear out all residents details
-          val result = controller.submit(
-            fakePostRequest.withFormUrlEncodedBody(
-              "answer" -> "no"
-            )
-          )
-          status(result) shouldBe SEE_OTHER
-          redirectLocation(result).value   shouldBe routes.HowManyNightsController.show.url
+          val result = controller.submit(fakePostRequest.withFormUrlEncodedBody("answer" -> "no"))
+          status(result)                    shouldBe SEE_OTHER
+          redirectLocation(result).value    shouldBe routes.HasCompletedLettingsController.show.url
           verify(repository, once).saveOrUpdate(data.capture())(any[Writes[Session]], any[HeaderCarrier])
-          hasCompletedLettings(data).value shouldBe false
-          completedLettings(data)          shouldBe Nil
+          hasPermanentResidents(data).value shouldBe false
+          permanentResidents(data)          shouldBe Nil
         }
       }
       "and the maximum number of residents has been reached" should {
-        "be handling POST hasCompletedLettings='yes' and reply 303 redirect to the 'Occupiers List' page" in new ControllerFixture(
-          hasCompletedLettings = Some(true)
+        "be handling POST answer='yes' by replying 303 redirect to the 'Resident List' page" in new ControllerFixture(
+          fiveResidents
         ) {
-          pending
           val result = controller.submit(fakePostRequest.withFormUrlEncodedBody("answer" -> "yes"))
-          status(result)                   shouldBe SEE_OTHER
-          redirectLocation(result).value   shouldBe routes.OccupierListController.show.url
+          status(result)                    shouldBe SEE_OTHER
+          redirectLocation(result).value    shouldBe routes.ResidentListController.show.url
           verify(repository, once).saveOrUpdate(data.capture())(any[Writes[Session]], any[HeaderCarrier])
-          hasCompletedLettings(data).value shouldBe true
-          completedLettings(data)            should have size 5
+          hasPermanentResidents(data).value shouldBe true
+          permanentResidents(data)            should have size 5
         }
       }
     }
   }
 
-  trait ControllerFixture(permanentResidents: List[ResidentDetail] = Nil, hasCompletedLettings: Option[Boolean] = None)
+  trait ControllerFixture(permanentResidents: List[ResidentDetail] = Nil)
       extends MockRepositoryFixture
       with SessionCapturingFixture:
-    val controller = new CompletedLettingsController(
+    val controller = new HasPermanentResidentsController(
       mcc = stubMessagesControllerComponents(),
       navigator = inject[LettingHistoryNavigator],
-      theView = inject[CompletedLettingsView],
+      theView = inject[HasPermanentResidentsView],
       sessionRefiner = preEnrichedActionRefiner(
         lettingHistory = Some(
           LettingHistory(
             hasPermanentResidents = Some(permanentResidents.nonEmpty),
-            permanentResidents = permanentResidents,
-            hasCompletedLettings = hasCompletedLettings
+            permanentResidents = permanentResidents
           )
         )
       ),
